fix: limit switch logic and converting rotations to meters
The limit switch logic for the elevator, making it have limits was fixed and the correct factor for motor encoder rotation to meter was used. The bottom limit switch logic just used to be if (not top limitswitch was true) which doesnt make sense. Breaking change: Possible wrong way to convert enncoder rotations to meters.